A sample character vector containing the background gene list used to perform FEA on the macrophage dataset

A sample character vector containing the background gene list used to perform FEA on the macrophage dataset

Format

A character vector

Value

A sample character vector containing the assay's rownames of the macrophage data

Details

This character vector object contains the assay's rownames of the macrophage data

The code to create said object can be found in the folder /inst/scripts in the EMMA package, the file is called create_datasets_examples.R.

References

Alasoo, et al. "Shared genetic effects on chromatin and gene expression indicate a role for enhancer priming in immune response", Nature Genetics, January 2018 doi: 10.1038/s41588-018-0046-7.